requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Environment
python-dotenv==1.0.0

# AI Integration
openai==1.3.7

# API Documentation
drf-spectacular==0.26.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Development
gunicorn==21.2.0